+++ changelog.md
@@ -1,17 +1,23 @@
 # Changelog
+
 All notable changes to the BSPED Paediatric DKA Calculator codebase will be documented in this file.
 
 ## [v2.0] - TBC 2024
+
 ### Changed
+
 - Total client-side codebase rebuild moving from Vue 2 to Vue 3 Composition with build step. No longer supporting internet explorer. The new build is a single page application with various user interface improvements.
 - Calculations of variables for care pathway have now moved to a server-side API call in place of the client-side script. Generation of the care pathway document is still done client-side.
 - DKA severity grade selection has changed to include consideration of bicarbonate level. If provided, a severity level will be selected based on both pH and bicarbonate with the more severe option used if these differ.
+
 ### Added
+
 - A unique patient hash is generated and stored using SHA-256 (x1 client-side, x1 server-side with salt) using patient NHS number and date of birth as input string. This allows episodes relating to the same patient to be linked for audit purposes.
 - The patient postcode is collected and sent with other data to the API where an index of multiple deprivation (IMD) decile is derived and stored. The postcode is not stored.
 - Bicarbonate, glucose and ketones may now optionally be entered and will be printed on the care pathway if provided. Bicarbonate is used for severity scoring as above.
 - Previously a care pathway could be generated even if the diagnostic thresholds for DKA were not met. A check to prevent this has been added. Ketones, if provided, must be above 3 mmol/L. pH must be <= 7.3 or bicarbonate must be < 15 mmol/L. If these thresholds are not met the API will return an error which must be corrected before a care pathway can be generated.
 - The patient ethnic group is now collected for audit purposes.
+- The insulin delivery method the patient uses (if they have pre-existing diabetes) is now collected for audit purposes.
 - Factors that may have contributed to the episode of DKA occuring (preventable factors) are now collected for audit purposes.
 
 ## [v1.3.5] - 2022-12-06 08:55